Move table conversion back to the import utils
